requirements.txt
openai
azure-search-documents
python-dotenv
psycopg[binary]>=3.1
